feat(ref_geo): removed ref_geo dependency
Replaced by an API call to get the municipality from a latitude
and a longitude. Called each time an observation is added.
Make this URL a parameter in the configuration
Updated the front to get only the name of the municipality
In the DataBase the municipality is the name of the town/city