fix(ref_geo): Catch exception when getting city
Added try/except to catch errors in get_municipality function to
prevent new observation from being created.
The main encountered error is a DNS error caused by WSL + Docker
apparently. Solved by restarting docker service. For this to never
happen, this call can be done in the frontend part and the city
can be sent to the backend