feat(city): add call to API_CITY in front
Since there can be problems in calling nominatim (API_CITY) in the
backend. It is now also called in the frontend with the same
API_CITY parameter in the app.config.
The API_CITY is called in the backend if the frontend does not
manage to get a city (municipality)